% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_cycle_processes.R
\name{all_individuals}
\alias{all_individuals}
\title{Life tables per age and maturation age}
\usage{
all_individuals(stock, fishing)
}
\arguments{
\item{stock}{Fish life-history parameters}

\item{fishing}{Table-defined fishing mortality}
}
\description{
Life tables per age and maturation age
}
