% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{dynamics}
\alias{dynamics}
\title{Fish stock dynamics}
\usage{
dynamics(stock, fishing, ssb, stock_dynamics)
}
\arguments{
\item{stock}{Initial life-history parameters}

\item{fishing}{Table-defined fishing mortality}

\item{ssb}{Initial SSB value}

\item{stock_dynamics}{Trajectory of changes in life-history parameters}
}
\value{
Trajectory of changes in abundance, biomass, average age and 
body length (per group)
}
\description{
Population dynamics is determined by the spawner-recruit relationship and 
transition probabilities between classes corresponding to elements in the 
Leslie matrix. Changing life-history parameters alter spawner-recruit 
relationship and transition probabilities. Initial population distribution is
calibrated to the observed spawning stock biomass (SSB).
}
