% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheryLHM.R
\docType{package}
\name{fisheryLHM-package}
\alias{fisheryLHM}
\alias{fisheryLHM-package}
\title{fisheryLHM: Fish stock life history model}
\description{
This package implements a bioenergetically driven age-, size-, and 
stage-specific model of a fish stock (Matsumura et al. 2011).
}
\references{
\itemize{
  \item Matsumura, S., Arlinghaus, R., & Dieckmann, U. (2011). Assessing 
  evolutionary consequences of size-selective recreational fishing on 
  multiple life-history traits, with an application to northern pike (Esox 
  lucius). Evolutionary Ecology, 25: 711-735. 
  https://doi.org/10.1007/s10682-010-9444-8
}
}
\keyword{internal}
