% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_history_model.R
\name{juvenile_recruitment}
\alias{juvenile_recruitment}
\title{Recruitment model for transitions from spawning stock biomass (SSB) to 
juveniles age 1}
\usage{
juvenile_recruitment(stock, life_history, spawner_recruit_model)
}
\arguments{
\item{stock}{Fish Life-history parameters}

\item{life_history}{Life tables}

\item{spawner_recruit_model}{Initial spawner-recruit model}
}
\value{
Spawner-juvenile model
}
\description{
Recruitment model for transitions from spawning stock biomass (SSB) to 
juveniles age 1
}
