% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{properties_at_equilibrium}
\alias{properties_at_equilibrium}
\title{Properties of demographic equilibrium}
\usage{
properties_at_equilibrium(stock, fishing)
}
\arguments{
\item{stock}{Fish life-history parameters}

\item{fishing}{Table-defined fishing mortality}
}
\value{
Abundance, biomass, average age and body length per group
}
\description{
The cases of density-independent reproduction (Leslie model) and density-
dependent reproduction (Beverton-Holt and Ricker models) are supported.
}
